requirements.txt
streamlit
opencv-python-headless
easyocr
inference-sdk
numpy
pillow
torch
torchvision